backend/requirements.txt
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
Bottleneck==1.3.7
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
dotenv==0.9.9
fastapi==0.110.0
ffmpy==0.6.0
filelock==3.18.0
fonttools==4.58.4
fsspec==2025.5.1
gradio==4.44.1
gradio_client==1.3.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.27.0
huggingface-hub==0.33.1
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.4
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mdurl==0.1.2
narwhals==1.44.0
numexpr==2.8.7
numpy==1.26.4
openai==1.30.1
orjson==3.10.1
packaging==25.0
pandas==2.2.2
pillow==10.4.0
pyarrow==15.0.2
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.9
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rich==14.0.0
rpds-py==0.25.1
ruff==0.12.0
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.36.3
tomlkit==0.12.0
tqdm==4.67.1
typer==0.16.0
typing_extensions==4.11.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.29.0
websockets==11.0.3
